update: modify procedure section layout and posts file structure, add done img
