+++ src/Components/landLordManagement/posts/procedures/BasicInformation.tsx
@@ -1,8 +1,8 @@
 import { useContext } from "react";
 import { useForm } from "react-hook-form";
-import cities from "../../../constants/locations/cities";
-import houseTypes from "../../../constants/houseTypes";
-import { ProcedureContext } from "../../../pages/landlordManagement/AddNew";
+import cities from "../../../../constants/locations/cities";
+import houseTypes from "../../../../constants/houseTypes";
+import { ProcedureContext } from "../../../../pages/landlordManagement/AddNew";
 
 export default function BasicInformation() {
   const { register, handleSubmit, watch } = useForm({
